requirements.txt
Django==1.10.3
djangorestframework==3.5.2
jsonfield==1.0.3
celery==4.0.2
psycopg2==2.6.2
channels==1.0.3
asgi_redis==1.0.0
daphne==1.0.3
flower==0.9.1
SharedArray==2.0.2
gunicorn
celery[redis]
git+ssh://git@github.com/KDD-OpenSource/predots-python-hics.git@0.7

# Disable these for docker envs
# scipy
# pandas
# numpy

# Testing
pytest-django==3.0.0
factory-boy
django-extensions
